requirements-to-freeze.txt
# ==============================
# Web server
# ==============================

Django<1.10,>1.9
gevent
gunicorn
dj_static



# ==============================
# Background processing
# ==============================

celery<4.0  # For Django 1.8
django-celery



# ==============================
# Database and Persistence
# ==============================

# DB Settings and Management
psycopg2-binary
psycogreen
dj-database-url
django-object-actions

# Caching
django-redis<4.9.0  # For Django 1.8
django-cache-url

# File Storage
boto
django-storages<1.7  # For Django 1.8



# ==============================
# User Accounts and Social Media
# ==============================

django-loginas
social-auth-core
social-auth-app-django
django-cors-headers<2.5.0  # For Django 1.8



# ==============================
# REST API
# ==============================

djangorestframework>3.3,<3.4
djangorestframework-csv==1.3.0
git+https://github.com/mjumbewu/django-rest-framework-bulk.git@492d76237dca99373d958daa68c9b3641bc7bbc5#egg=djangorestframework-bulk==0.2.1-m1
six
markdown<3  # For browsable API docs
python-dateutil
ujson
Pillow

# The Django admin interface
django-ace<1.0.4  # For Django 1.8

# The manager interface
requests



# ==============================
# Testing, validating, and debugging
# ==============================

nose
django-nose
mock
mock_django
responses
django-debug-toolbar==1.2.1
raven  # For Sentry error logging


# - - - - - - - - - - - - - - - -

# For DRF 0.4 (deprecated)
URLObject>=0.6.0
